// web/src/types/user/rbac/user-role.types.ts
/*
 * Copyright (c) 2026 lovelycat
 *
 * This source code is licensed under the MIT license found in the
 * LICENSE file in the root directory of this source tree.
 */

import type {BaseEntity} from "../../BaseEntity.ts";

export interface UserRole extends BaseEntity {
    name: string;
    description: string | null;
}